+++ tests/TestCase.php
@@ -4,6 +4,7 @@ namespace DoubleThreeDigital\Runway\Tests;
 
 use DoubleThreeDigital\Runway\Routing\Traits\RunwayRoutes;
 use DoubleThreeDigital\Runway\ServiceProvider;
+use DoubleThreeDigital\Runway\Traits\HasRunwayResource;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Encryption\Encrypter;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
@@ -127,6 +128,13 @@ abstract class TestCase extends OrchestraTestCase
                                         ],
                                     ],
                                     [
+                                        'handle' => 'excerpt',
+                                        'field' => [
+                                            'type' => 'textarea',
+                                            'read_only' => true,
+                                        ],
+                                    ],
+                                    [
                                         'handle' => 'author_id',
                                         'field' => [
                                             'type' => 'belongs_to',
@@ -232,10 +240,14 @@ abstract class TestCase extends OrchestraTestCase
 
 class Post extends Model
 {
-    use RunwayRoutes;
+    use RunwayRoutes, HasRunwayResource;
 
     protected $fillable = [
         'title', 'slug', 'body', 'author_id',
+    ];
+
+    protected $appends = [
+        'excerpt',
     ];
 
     public function scopeFood($query)
@@ -254,10 +266,17 @@ class Post extends Model
     {
         return $this->belongsTo(Author::class);
     }
+
+    public function getExcerptAttribute()
+    {
+        return 'This is an excerpt.';
+    }
 }
 
 class Author extends Model
 {
+    use HasRunwayResource;
+
     protected $fillable = [
         'name',
     ];