+++ CHANGELOG.md
@@ -6,6 +6,26 @@
 
 -   Runway v5 now supports Statamic 4 #235
 
+## v4.2.4 (2023-04-15)
+
+### What's fixed
+
+-   Slug field will be used for `slug` column when generating blueprint from Eloquent model #251
+
+## v4.2.3 (2023-04-08)
+
+### What's fixed
+
+-   Fixed "Column index out of range" error #242
+-   Various fixes to Table mode on the Has Many fieldtype #246
+-   Fixed an issue where Runway would try to save attributes from the `$appends` array #247
+-   Fixed an 'ambiguous table' error with one of Runway's queries #248
+-   Fix double requests, caused by column data changing behind the scenes #250
+
+### What's improved
+
+-   Breadcrumbs are passed into the inline publish form (the one that shows when you edit something in a stack) #249
+
 ## v4.2.2 (2023-02-28)
 
 ### What's fixed